custom from_params config lives with from_params method
